+++ cmd/aperturectl/cmd/blueprints/root.go
@@ -15,7 +15,7 @@ import (
 )
 
 const (
-	defaultBlueprintsRepo    = "github.com/fluxninja/aperture/v2/blueprints"
+	defaultBlueprintsRepo    = "github.com/fluxninja/aperture/blueprints"
 	defaultBlueprintsVersion = "latest"
 )
 
@@ -43,6 +43,7 @@ func init() {
 	BlueprintsCmd.AddCommand(removeCmd)
 	BlueprintsCmd.AddCommand(generateCmd)
 	BlueprintsCmd.AddCommand(valuesCmd)
+	BlueprintsCmd.AddCommand(dynamicValuesCmd)
 }
 
 // BlueprintsCmd is the root command for blueprints.
